Fix XTENLANG-3418: use Long instead of Int in Reader.available and mark
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@28001 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/core/io/InputStream.java
@@ -103,18 +103,18 @@ public class InputStream extends Ref {
         }
     }
     
-    public int available() {
+    public long available() {
         try {
-            return stream.available();
+            return (long) stream.available();
         } catch (java.io.IOException e) {
             throw new x10.io.IOException(e.getMessage());
         }
     }
     
     // XTENLANG-2680
-    public int available$O() {
+    public long available$O() {
         try {
-            return stream.available();
+            return (long) stream.available();
         } catch (java.io.IOException e) {
             throw new x10.io.IOException(e.getMessage());
         }
@@ -128,8 +128,9 @@ public class InputStream extends Ref {
         }
     }
     
-    public void mark(int readlimit) {
-        stream.mark(readlimit);
+    // LONG_RAIL: unsafe int cast
+    public void mark(long readlimit) {
+        stream.mark((int) readlimit);
     }
     
     public void reset() {
